Restructuring classes to internals
