+++ clients/src/main/java/org/apache/kafka/common/telemetry/internals/MetricKey.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.common.telemetry.metrics;
+package org.apache.kafka.common.telemetry.internals;
 
 import org.apache.kafka.common.MetricName;
