+++ clients/src/main/java/org/apache/kafka/common/telemetry/internals/MetricsCollector.java
@@ -14,9 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.common.telemetry.collector;
-
-import org.apache.kafka.common.telemetry.emitter.Emitter;
+package org.apache.kafka.common.telemetry.internals;
 
 /**
  * A {@code MetricsCollector} is responsible for scraping a source of metrics and forwarding
