+++ clients/src/main/java/org/apache/kafka/common/telemetry/internals/SinglePointMetric.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.common.telemetry.metrics;
+package org.apache.kafka.common.telemetry.internals;
 
 /**
  * This class represents a metric that does not yet contain resource tags.
